feat(data context) Install data context as a service
Install the data context as a service with the name 'data' as well as each
key in the data context. Each installed service will be a ReactiveVar
instance that will update when the data context changes. This was done so
that data form the data context can be depended on just like any other
dependency.